refactor: 불필요한 클래스 제거 및 폴더 이동
