+++ README.md
@@ -68,7 +68,6 @@
 
 - [x] 1단계 피드백 내용 적용
     - [x] 토큰 확인 할 때 현재 가입된 email 인가?에 대한 검증을 interceptor에서도 해주자.
-    - [ ] 도메인 단에서도 검증을 이루어질 수 있도록 변경한다 ( 구현까지 시간이 촉박하므로 마지막 이슈로... )
 - [x] 모든 상품 조회 API 구현
     - 상품 조회 시에는 `GET` `/api/products` 로 요청한다.
     - List를 이용해 모든 상품이 조회된다.
@@ -94,3 +93,4 @@
     - `예외` count가 물품의 재고를 넘기는 경우
     - `예외` 존재하지 않는 productId로 요청을 보낼 경우
     - `예외` customerId와 입력받은 토큰의 유저와 다를 경우
+- [ ] 도메인 단에서도 검증을 이루어질 수 있도록 변경한다 ( 구현까지 시간이 촉박하므로 마지막 이슈로... )
