+++ src/main/java/woowacourse/shoppingcart/dto/customer/CustomerResponse.java
@@ -1,6 +1,6 @@
 package woowacourse.shoppingcart.dto.customer;
 
-import woowacourse.shoppingcart.domain.Customer;
+import woowacourse.shoppingcart.domain.customer.Customer;
 
 public class CustomerResponse {
